Refactor AB3 terms in Test files.
